fix: ensure we can publish to crates
include_str won't work outside of the crate boundary. Crates is so
frustrating sometimes.
